requirements.txt
openai-whisper
torchaudio
soundfile
numpy
torch